refactor: GithubRepoListContainer, GithubRepoItem 컴포넌트 분리
